Update stsrun_sqlite.py
Added functionality to find latest run in the database and download the missing runs in a zip file from baalorlord.tv, extract it, then load those files and remove them. This allows for automation.

If database is empty, it defaults to the minimum timestamp for runs that can be loaded.